🔃Rename TextChunker from SemanticTextPartitioner
This commit refactors the SemanticTextPartitioner class, which splits text into chunks based on punctuation and new lines, and renames it to TextChunker. The new name better reflects the functionality of the class and avoids confusion with the SemanticText and SemanticPartitioner classes, which are different abstractions for text analysis. The TextChunker class is also moved from the SemanticFunctions.Partitioning namespace to the Text namespace, as it is a utility class rather than a semantic function. The commit updates the references to the class in the ConversationSummarySkill and the DocumentImportController classes, as well as the documentation.
