+++ dotnet/src/SemanticKernel/Text/FunctionExtensions.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.SemanticKernel.Orchestration;
 using Microsoft.SemanticKernel.SkillDefinition;
 
-namespace Microsoft.SemanticKernel.SemanticFunctions.Partitioning;
+namespace Microsoft.SemanticKernel.Text;
 
 /// <summary>
 /// Class with extension methods for semantic functions.
